pyproject.toml
[project]
name = "wb-mqtt-topic-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt==2.1.0",
    "pytest==9.0.1",
    "pytest-asyncio==1.3.0",
    "setuptools==80.9.0",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "N", "YTT", "B", "A", "DJ", "T20", "SIM", "ERA", "I001"]
ignore = ["SIM102", "B008", "SIM114", "A003", "C901"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode="auto"
